重症患者数: Rename components/SevereCaseBarChart.vue to components/index/CardsMonitoring/SevereCase/Chart.vue
