+++ components/index/CardsMonitoring/SevereCase/Chart.vue
